Specify subcommand list in subcommand.rs
This is superior to having it both in main.rs AND argparse.rs which was
the previous solution. Passing the subcommands to find_tags_in_args
might be a bit more complex than just calling get_subcommands() itself
in the function. Might be a better/simpler solution, but is it less
performant(?)