Assume default branch if no branch was found with `symbolic-ref`
If we can't find the default branch with the `git symbolic-ref
refs/remotes/origin/HEAD` we assume default branch is `master`.
The value that is assumed is stored in a const for easy change in the
future (main?).

We might consider assuming both `master` and `main` as defaults as they
are only used to highlight the output of the `status` command.

Fetching the actual default branch with `git remote show <remote>` is
way to slow, and not worth it.