Add `tmux` subcommand for initial support for launching tmux
Launching a tmux session with all specified dirs is practical in many
ways. This is the initial support, where we simply spawn a new session
and split the initial window into n number of panes.

In the future we would probably like to add arguments to the `tmux`
subcommand that allows us to specify layout modes and whether to use
windows or panes.